fix dev TLD
I'd like to take this opportunity to say that not only did ICANN violate
the spirit of their bylaws by allowing the .dev TLD to be allocated at
all, but that Google violated the terms of the registry operating
charter they published which stated the explicit purpose would only ever
be internal use of the .dev TLD.

Further, a simple search for existing usage of the TLD would have
revealed extensive use of the .dev TLD in various development tools for
testing services internally such as Pow and others. The team that did
the due diligence utterly failed to perform even the bare minimum of
effort required or expected to maintain the integrity of the DNS system.